autograder/requirements.txt
Flask==2.1.1
Pillow==9.1.0  
pdf2image==1.16.0  
pytesseract==0.3.8  
opencv-python==4.5.5.64
spellchecker==0.4.0  
scikit-learn==1.0.2  
textblob==0.15.3 
werkzeug==2.1.1  

 <button onclick="window.location.href='http://localhost:5000/'" class="home">
        Home
      </button>